+++ spec/truffle/interop/foreign_string_spec.rb
@@ -26,8 +26,28 @@ describe "Foreign strings" do
   
   it "are unboxed and converted to Ruby automatically on the RHS of string concatenation" do
     a = 'a'
-    b = Truffle::Debug.foreign_string('a')
+    b = Truffle::Debug.foreign_string('b')
     (a + b).should == 'ab'
+  end
+  
+  it "respond to #to_s" do
+    Truffle::Debug.foreign_string('a').to_s.should == 'a'
+  end
+  
+  it "respond to #to_str" do
+    Truffle::Debug.foreign_string('a').to_str.should == 'a'
+  end
+  
+  it "can be printed with #puts" do
+    lambda {
+      puts Truffle::Debug.foreign_string('a')
+    }.should output("a\n")
+  end
+  
+  it "can be printed as if a Ruby string with #p" do
+    lambda {
+      p Truffle::Debug.foreign_string('a')
+    }.should output("\"a\"\n")
   end
 
 end