+++ src/main/ruby/core/truffle/interop.rb
@@ -169,13 +169,26 @@ module Truffle
         else
           Truffle::Interop.invoke(receiver, :class, *args)
         end
+      when :inspect
+        if Truffle::Interop.size?(receiver)
+          "#<Foreign #{to_array(receiver).inspect}>"
+        elsif Truffle::Interop.boxed?(receiver)
+          Truffle::Interop.unbox(receiver).inspect
+        else
+          "#<Foreign:0x#{Truffle::Interop.identity_hash_code(receiver).to_s(16)}>"
+        end
       when :to_s
-        Truffle::Interop.to_string(receiver)
+        receiver = Truffle::Interop.unbox_if_needed(receiver)
+        if receiver.is_a?(String)
+          receiver
+        else
+          receiver.inspect
+        end
       when :to_str
         receiver = Truffle::Interop.unbox_if_needed(receiver)
         raise NameError, 'no method to_str' unless receiver.is_a?(String)
         receiver
-      when :is_a?
+      when :is_a?, :kind_of?
         receiver = Truffle::Interop.unbox_if_needed(receiver)
         check_class = args.first
         if Truffle::Interop.foreign?(receiver)
@@ -215,7 +228,7 @@ module Truffle
       when :to_str
         object = Truffle::Interop.unbox_if_needed(object)
         !Truffle::Interop.foreign?(object) && object.is_a?(String)
-      when :inspect, :to_s, :is_a?
+      when :inspect, :to_s, :is_a?, :kind_of?
         true
       else
         false